moved performance metrics to its own componenet for reusability
